apps/api-core/requirements.txt
# Production dependencies
# Alternative to Poetry - use this if not using Poetry for dependency management
# To install: pip install -r requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0  # Pin to 4.3.0 to avoid compatibility issues with bcrypt 5.0.0
httpx==0.25.2
email-validator==2.1.0
azure-storage-blob==12.19.0  # Azure Blob Storage SDK for file uploads
python-multipart==0.0.6  # Required for FastAPI file uploads
aio-pika==9.2.1  # RabbitMQ publishing (document ingestion jobs)
